+++ strutil/example_test.go
@@ -9,7 +9,6 @@ package strutil
 
 import (
 	"fmt"
-	"strconv"
 	"strings"
 )
 
@@ -235,7 +234,7 @@ func ExampleMask() {
 func ExampleJoinFunc() {
 	e := []string{"John", "Bob Smith", "Fiona", "Elen Vire", "Joe Doe"}
 	s := JoinFunc(e, ", ", func(s string) string {
-		return strconv.Quote(s)
+		return `"` + s + `"`
 	})
 
 	fmt.Println(s)